Fix compiler errors.

+++ src/pipeline/test/RadeTextTest.cpp
@@ -24,9 +24,9 @@ void addNoise(short* ptr, int size)
     for (int i = 0; i < size; i++)
     {
         double noise = 16384 * distribution(generator);
-        noise = (noise + ptr[i]) / 0.707;
-        if (noise > INT_MAX) noise = INT_MAX;
-        else if (noise < INT_MIN) noise = INT_MIN;
+        noise = (noise + (double)ptr[i]) / 0.707;
+        if (noise > SHRT_MAX) noise = SHRT_MAX;
+        else if (noise < SHRT_MIN) noise = SHRT_MIN;
         ptr[i] = noise;
     }
 }
